move vk-regression to /tests
